# shared/openapi/mesob-api.yaml
openapi: "3.1.0"
info:
  title: Mesob Wallet API
  version: "1.0.0"
  description: |
    Unified public API for the Mesob Wallet platform.
    All requests go through the API Gateway at port 8000.

    ## Authentication
    - Mobile/USSD users authenticate with **MSISDN + PIN** via `/v1/identity/login`
    - Access tokens are RS256 JWTs valid for **15 minutes**
    - Refresh tokens are 7-day opaque tokens returned as an `httpOnly` cookie
    - Staff users (ADMIN, BRANCH_MANAGER, SUPER_ADMIN) use the same auth flow

    ## Demo Accounts
    | Role | MSISDN | PIN |
    |------|--------|-----|
    | SUPER_ADMIN | +251911000001 | 111111 |
    | ADMIN | +251911000002 | 111111 |
    | BRANCH_MANAGER | +251911000003 | 111111 |
    | AGENT | +251911000004 | 111111 |
    | USER (KYC Tier 2) | +251911000005 | 111111 |
    | USER (KYC Tier 0) | +251911000006 | 111111 |

    ## Amounts
    All monetary values are in **ETB minor units (cents)**.
    Example: 100.00 ETB = `10000`

    ## Response Envelope
    All responses follow:
    ```json
    {
      "data": { ... },
      "meta": { "request_id": "...", "trace_id": "..." }
    }
    ```
    Errors:
    ```json
    {
      "error": { "code": "DOMAIN_ERROR_CODE", "message": "human message" },
      "meta": { ... }
    }
    ```
  contact:
    name: Mesob Engineering
    email: engineering@mesobwallet.et
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.mesobwallet.et/v1
...
paths:
...
  /iqub/groups:
    post:
      tags: [Iqub]
      summary: Create a rotating savings group
      operationId: createIqubGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IqubGroupCreate"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    get:
      tags: [Iqub]
      summary: List iqub groups the caller belongs to
      operationId: listIqubGroups
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of iqub groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /iqub/groups/{id}:
    get:
      tags: [Iqub]
      summary: Get iqub group detail
      operationId: getIqubGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Group detail
        "404":
          $ref: "#/components/responses/NotFound"
  /iqub/groups/{id}/contribute:
    post:
      tags: [Iqub]
      summary: Make cycle contribution
      operationId: contributeIqub
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idempotency_key]
              properties:
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Contribution posted
        "402":
          description: Insufficient balance
        "409":
          description: Already contributed this cycle

  # ── IDDIR ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT access token. Obtain via /identity/login or /identity/token/refresh.
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        data:
          description: Response payload (schema varies by endpoint)
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [data]
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INSUFFICIENT_BALANCE"
            message:
              type: string
              description: Human-readable message
              example: "Your balance is too low for this transaction"
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [error]
    MetaBlock:
      type: object
      properties:
        request_id:
          type: string
          example: "req_01J3ABCDEF"
        trace_id:
          type: string
          example: "trace_01J3ABCDEF"
    IqubGroupCreate:
      type: object
      required: [name, cycle_amount_minor, cycle_days, member_msisdns]
      properties:
        name:
          type: string
          maxLength: 128
          example: "Office Iqub 2026"
        cycle_amount_minor:
          type: integer
          format: int64
          description: Amount each member contributes per cycle (ETB cents)
          example: 50000
        cycle_days:
          type: integer
          description: Days per cycle (e.g. 30 = monthly)
          example: 30
        member_msisdns:
          type: array
          minItems: 2
          maxItems: 20
          items:
            type: string
          example: ["+251911111111", "+251922222222", "+251933333333"]
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
...